third_party/blink/web_tests/external/wpt/css/selectors/is-where-error-recovery-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Selectors: :is() and :where() error recovery
  assert_equals: Should be parsed as-is (but not be considered valid) expected ":is(:total-nonsense)" but got ":is()"
Harness: the test ran to completion.
